+++ src/main/java/net/lopymine/pe/entrypoint/PEFabricEntrypoint.java
@@ -2,7 +2,7 @@ package net.lopymine.pe.entrypoint;
 
 //? if fabric {
 
-import net.lopymine.pe.ParticleEffects;
+/*import net.lopymine.pe.ParticleEffects;
 
 import net.fabricmc.api.ModInitializer;
 
@@ -14,4 +14,4 @@ public class PEFabricEntrypoint implements ModInitializer {
 	}
 }
 
-//?}
+*///?}