+++ CHANGELOG.md
@@ -1,1 +1,2 @@
-- Ported to 26.1
+- Fixed compatibility with new version (v2.0.1) of Inventory Particles
+- Ported to 26.1 (NeoForge)